move state2 to state
