+++ lib/rktk-keymanager/src/lib.rs
@@ -17,7 +17,6 @@ pub mod interface;
 pub mod keycode;
 pub mod keymap;
 mod macros;
-pub mod state2;
 mod time;
 
 #[cfg(any(test, feature = "state"))]
