+++ lib/rktk-keymanager/src/state/hid_report.rs
