+++ lib/rktk-keymanager/src/state/tests/action/tap_dance.rs
@@ -1,4 +1,4 @@
-use crate::state::tests::prelude::*;
+use super::super::prelude::*;
 use pretty_assertions::assert_eq;
 
 #[test]
