--- lib/rktk-keymanager/src/state2/shared.rs
@@ -1,74 +1,0 @@
-use crate::time::{Duration, Instant};
-
-use crate::keymap::Keymap;
-
-pub(super) type LayerActive<const LAYER: usize> = [bool; LAYER];
-
-pub(super) struct SharedState<
-    const LAYER: usize,
-    const ROW: usize,
-    const COL: usize,
-    const ENCODER_COUNT: usize,
-    const TAP_DANCE_MAX_DEFINITIONS: usize,
-    const TAP_DANCE_MAX_REPEATS: usize,
-    const COMBO_KEY_MAX_DEFINITIONS: usize,
-    const COMBO_KEY_MAX_SOURCES: usize,
-> {
-    pub keymap: Keymap<
-        LAYER,
-        ROW,
-        COL,
-        ENCODER_COUNT,
-        TAP_DANCE_MAX_DEFINITIONS,
-        TAP_DANCE_MAX_REPEATS,
-        COMBO_KEY_MAX_DEFINITIONS,
-        COMBO_KEY_MAX_SOURCES,
-    >,
-    pub layer_active: LayerActive<LAYER>,
-    pub now: Instant,
-}
-
-impl<
-        const LAYER: usize,
-        const ROW: usize,
-        const COL: usize,
-        const ENCODER_COUNT: usize,
-        const TAP_DANCE_MAX_DEFINITIONS: usize,
-        const TAP_DANCE_MAX_REPEATS: usize,
-        const COMBO_KEY_MAX_DEFINITIONS: usize,
-        const COMBO_KEY_MAX_SOURCES: usize,
-    >
-    SharedState<
-        LAYER,
-        ROW,
-        COL,
-        ENCODER_COUNT,
-        TAP_DANCE_MAX_DEFINITIONS,
-        TAP_DANCE_MAX_REPEATS,
-        COMBO_KEY_MAX_DEFINITIONS,
-        COMBO_KEY_MAX_SOURCES,
-    >
-{
-    pub fn new(
-        keymap: Keymap<
-            LAYER,
-            ROW,
-            COL,
-            ENCODER_COUNT,
-            TAP_DANCE_MAX_DEFINITIONS,
-            TAP_DANCE_MAX_REPEATS,
-            COMBO_KEY_MAX_DEFINITIONS,
-            COMBO_KEY_MAX_SOURCES,
-        >,
-    ) -> Self {
-        Self {
-            keymap,
-            layer_active: [false; LAYER],
-            now: Instant::from_start(Duration::from_millis(0)),
-        }
-    }
-
-    pub fn highest_layer(&self) -> usize {
-        self.layer_active.iter().rposition(|&x| x).unwrap_or(0)
-    }
-}
